+++ hata/discord/exceptions.py
@@ -362,6 +362,8 @@ class ERROR_CODES:
     +---------------------------------------------------+-----------+-------+
     | unknown_application_command_permissions           | 10066     | -     |
     +---------------------------------------------------+-----------+-------+
+    | unknown_stage                                     | 10067     | -     |
+    +---------------------------------------------------+-----------+-------+
     | bots_not_allowed                                  | 20001     | -     |
     +---------------------------------------------------+-----------+-------+
     | only_bots_allowed                                 | 20002     | -     |
@@ -590,6 +592,7 @@ class ERROR_CODES:
     unknown_application_command = 10063
     unknown_voice_state = 10065
     unknown_application_command_permissions = 10066
+    unknown_stage = 10067
     
     bots_not_allowed = 20001
     only_bots_allowed = 20002